# swagger.yaml
openapi: 3.0.0
info:
  title: Incident Report API
  description: A production-ready RESTful API for building management with component tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: https://uncomprehensive-jeopardously-sherryl.ngrok-free.dev/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /buildings/{id}/floors:
    get:
      tags:
        - Buildings
        - Floors
      summary: Get Floors in Building
      description: Get all floors in a specific building
      operationId: getFloorsByBuilding
      parameters:
        - name: id
          in: path
          required: true
          description: Building ID
          schema:
            type: integer
      responses:
        '200':
          description: Floors retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloorsListResponse'
  /floors:
    post:
      tags:
        - Floors
      summary: Create Floor
      description: Create a new floor
      operationId: createFloor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFloorRequest'
      responses:
        '201':
          description: Floor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Floors
      summary: Get All Floors
      description: Get all floors with building information and pagination
      operationId: getAllFloors
      parameters:
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Records per page (default 10, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Floors retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFloorResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /floors/{id}:
    get:
      tags:
        - Floors
      summary: Get Floor by ID
      description: Retrieve a specific floor by ID
      operationId: getFloorById
      parameters:
        - name: id
          in: path
          required: true
          description: Floor ID
          schema:
            type: integer
      responses:
        '200':
          description: Floor retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloorResponse'
        '400':
          description: Floor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Floors
      summary: Update Floor
      description: Update a floor
      operationId: updateFloor
      parameters:
        - name: id
          in: path
          required: true
          description: Floor ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFloorRequest'
      responses:
        '200':
          description: Floor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloorResponse'
        '400':
          description: Update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Floors
      summary: Delete Floor
      description: Delete a floor
      operationId: deleteFloor
      parameters:
        - name: id
          in: path
          required: true
          description: Floor ID
          schema:
            type: integer
      responses:
        '200':
          description: Floor deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reports:
    post:
      tags:
        - Reports
      summary: Create Report
      description: Create a new report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Reports
      summary: Get All Reports
      description: Get all reports with pagination support
      operationId: getAllReports
      parameters:
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Records per page (default 10, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{id}:
    get:
      tags:
        - Reports
      summary: Get Report by ID
      description: Retrieve a specific report by ID
      operationId: getReportById
      parameters:
        - name: id
          in: path
          required: true
          description: Report ID
          schema:
            type: integer
      responses:
        '200':
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Reports
      summary: Update Report
      description: Update a report (partial update allowed)
      operationId: updateReport
      parameters:
        - name: id
          in: path
          required: true
          description: Report ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportRequest'
      responses:
        '200':
          description: Report updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Reports
      summary: Delete Report
      description: Delete a report from the system
      operationId: deleteReport
      parameters:
        - name: id
          in: path
          required: true
          description: Report ID
          schema:
            type: integer
      responses:
        '200':
          description: Report deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{id}/assign-user:
    put:
      tags:
        - Reports
      summary: Assign User to Report
      description: Assign a user to an existing report
      operationId: assignUserToReport
      parameters:
        - name: id
          in: path
          required: true
          description: Report ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUserRequest'
      responses:
        '200':
          description: User assigned to report successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Assignment failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    BuildingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Building created successfully
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            code:
              type: string
              example: BLDG001
            name:
              type: string
              example: Building A
            location:
              type: string
              example: 123 Main Street
            created_at:
              type: integer
              example: 1703000000
            updated_at:
              type: integer
              example: 1703000000
...
    CreateFloorRequest:
      type: object
      required:
        - building_id
        - number
        - name
      properties:
        building_id:
          type: integer
          example: 1
        floor_number:
          type: integer
          example: 0
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: Ground Floor
    UpdateFloorRequest:
      type: object
      properties:
        floor_number:
          type: integer
          example: 1
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: First Floor
    FloorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Floor created successfully
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            building_id:
              type: integer
              example: 1
            building:
              $ref: '#/components/schemas/BuildingResponse'
            floor_number:
              type: integer
              example: 0
            name:
              type: string
              example: Ground Floor
            created_at:
              type: integer
              example: 1703000000
            updated_at:
              type: integer
              example: 1703000000
    FloorsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Floors retrieved successfully
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              building_id:
                type: integer
              floor_number:
                type: integer
              name:
                type: string
              created_at:
                type: integer
              updated_at:
                type: integer
    PaginatedFloorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Floors retrieved successfully
        data:
          type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/FloorResponse'
            page:
              type: integer
              example: 1
            page_size:
              type: integer
              example: 10
            total:
              type: integer
              example: 2
            total_page:
              type: integer
              example: 1

    # Room Schemas
...
    CreateReportRequest:
      type: object
      required:
        - name
        - room_id
        - component_id
        - status
      properties:
        name:
          type: string
          example: Broken light fixture
        room_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        component_id:
          type: integer
          example: 1
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED]
          example: PENDING
    UpdateReportRequest:
      type: object
      properties:
        name:
          type: string
          example: Fixed light fixture
        room_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        component_id:
          type: integer
          example: 1
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED]
          example: COMPLETED
    AssignUserRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: integer
          example: 1
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Report created successfully
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Broken light fixture
            room_id:
              type: integer
              example: 1
            user_id:
              type: integer
              nullable: true
              example: 1
            component_id:
              type: integer
              example: 1
            status:
              type: string
              example: PENDING
            created_at:
              type: string
              format: date-time
              example: "2023-12-01T10:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-12-01T10:00:00Z"
    PaginatedReportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Reports retrieved successfully
        data:
          type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  room_id:
                    type: integer
                  user_id:
                    type: integer
                    nullable: true
                  component_id:
                    type: integer
                  status:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
            page:
              type: integer
              example: 1
            page_size:
              type: integer
              example: 10
            total:
              type: integer
              example: 2
            total_page:
              type: integer
              example: 1

    # Common Schemas
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Operation failed
        error:
          type: string
          example: Validation error details